[webgpu] Merge shader_preprocessor.ts into webgpu_program.ts
The merge can solve circular import problem so that WebGPUProgram can be
removed.
isFlatDispatch is added to avoid potential inconsist changes.
yarn.lock is changed due to structual change in tfjs root folder.
