+++ tfjs-backend-webgpu/src/batchnorm_webgpu.ts
@@ -16,9 +16,7 @@
  */
 
 import {backend_util} from '@tensorflow/tfjs-core';
-
-import {getMainHeaderAndGlobalIndexString} from './shader_preprocessor';
-import {WebGPUProgram} from './webgpu_program';
+import {getMainHeaderAndGlobalIndexString, WebGPUProgram} from './webgpu_program';
 import {computeDispatch, flatDispatchLayout} from './webgpu_util';
 
 export class BatchNormProgram implements WebGPUProgram {
