+++ tfjs-backend-webgpu/src/flip_left_right_webgpu.ts
@@ -15,10 +15,8 @@
  * =============================================================================
  */
 
-import {getMainHeaderAndGlobalIndexString} from './shader_preprocessor';
+import {getMainHeaderAndGlobalIndexString, WebGPUProgram} from './webgpu_program';
 import {computeDispatch, flatDispatchLayout} from './webgpu_util';
-
-import {WebGPUProgram} from './webgpu_program';
 
 export class FlipLeftRightProgram implements WebGPUProgram {
   outputShape: number[] = [];
