+++ tfjs-backend-webgpu/src/matmul_packed_webgpu.ts
@@ -16,10 +16,8 @@
  */
 
 import {backend_util, TensorInfo, util} from '@tensorflow/tfjs-core';
-
 import {mapActivationToShaderProgram} from './activation_util';
-import {getMainHeaderString} from './shader_preprocessor';
-import {WebGPUProgram} from './webgpu_program';
+import {getMainHeaderString, WebGPUProgram} from './webgpu_program';
 import {computeDispatch, computeWorkGroupSizeForMatMul} from './webgpu_util';
 
 const writeDataToSubASnippet = (transpose: boolean) => {
