+++ tfjs-backend-webgpu/src/resize_bilinear_webgpu.ts
@@ -15,8 +15,7 @@
  * =============================================================================
  */
 
-import {getMainHeaderAndGlobalIndexString} from './shader_preprocessor';
-import {WebGPUProgram} from './webgpu_program';
+import {getMainHeaderAndGlobalIndexString, WebGPUProgram} from './webgpu_program';
 import {computeDispatch, flatDispatchLayout} from './webgpu_util';
 
 export class ResizeBilinearProgram implements WebGPUProgram {
