+++ tfjs-backend-webgpu/src/scatter_optimized_webgpu.ts
@@ -16,9 +16,7 @@
  */
 
 import {DataType} from '@tensorflow/tfjs-core';
-
-import {getCoordsDataType, getMainHeaderAndGlobalIndexString} from './shader_preprocessor';
-import {WebGPUProgram} from './webgpu_program';
+import {getCoordsDataType, getMainHeaderAndGlobalIndexString, WebGPUProgram} from './webgpu_program';
 import {computeDispatch, flatDispatchLayout} from './webgpu_util';
 
 export class ScatterOptimizedProgram implements WebGPUProgram {
