+++ tfjs-backend-webgpu/src/top_k_webgpu.ts
@@ -15,8 +15,7 @@
  * =============================================================================
  */
 
-import {getMainHeaderAndGlobalIndexString} from './shader_preprocessor';
-import {WebGPUProgram} from './webgpu_program';
+import {getMainHeaderAndGlobalIndexString, WebGPUProgram} from './webgpu_program';
 import {computeDispatch, flatDispatchLayout} from './webgpu_util';
 
 // Based on Algorithm 2 of Bitonic Top K, ref:
