+++ tfjs-backend-webgpu/src/unary_op_webgpu.ts
@@ -15,11 +15,9 @@
  * =============================================================================
  */
 
-import {getMainHeaderAndGlobalIndexString} from './shader_preprocessor';
-import {computeDispatch, flatDispatchLayout} from './webgpu_util';
-
 import {getUnaryOpString, UnaryOpType} from './unary_op_util';
-import {WebGPUProgram} from './webgpu_program';
+import {getMainHeaderAndGlobalIndexString, WebGPUProgram} from './webgpu_program';
+import {computeDispatch, flatDispatchLayout} from './webgpu_util';
 
 export class UnaryOpProgram implements WebGPUProgram {
   outputShape: number[];
